+++ src/components/ClearButton/index.tsx
@@ -2,17 +2,10 @@ import { IfcClearButton } from '../../types';
 
 const ClearButton = (clearButtonProps: IfcClearButton) => {
   const { handleClear } = clearButtonProps;
-  const clearStatus = () => {
-    localStorage.removeItem('ElectoralCollegeStatus');
-    localStorage.removeItem('PopularVoteStatus');
-    localStorage.removeItem('PopularVoteTotals');
-    localStorage.removeItem('WinnerTakeAllTotals');
-    handleClear();
-  };
 
   return (
     <div className="buttonWrapper">
-      <button id="clearButton" type="button" onClick={clearStatus}>
+      <button id="clearButton" type="button" onClick={handleClear}>
         Clear
       </button>
     </div>